Added a unittest for VectorSpecies / VectorTransmission which _sometimes_ passes. Under investigation.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@421 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/Transmission/PerHost.h
@@ -127,7 +127,7 @@ private:
   int timestepITN;
   int timestepIRS;
   
-  int nextItnDistribution;
+  size_t nextItnDistribution;
   
   
   ///@brief Age-group variables for wtprop and ageSpecificRelativeAvailability
